# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * MOB HUB Egypt — Mobile Phone Marketplace API
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface StoreDto {
  id: string;
  nameAr: string;
  nameEn: string;
  descriptionAr: string;
  descriptionEn: string;
  logoColor: string;
  logoInitial: string;
  coverGradient: string[];
  rating: number;
  reviewsCount: number;
  governorate: string;
  city: string;
  address: string;
  addressAr: string;
  lat: number;
  lng: number;
  phone: string;
  whatsapp: string;
  facebook?: string | null;
  instagram?: string | null;
  website?: string | null;
  workingHours: string;
  workingHoursAr: string;
  isOpen: boolean;
  isVerified: boolean;
  productsCount: number;
  categories: string[];
}

export interface ProductDto {
  id: string;
  storeId: string;
  nameAr: string;
  nameEn: string;
  descriptionAr: string;
  descriptionEn: string;
  brand: string;
  model: string;
  price: number;
  discountPrice?: number | null;
  category: string;
  condition: string;
  inStock: boolean;
  warranty: string;
  warrantyAr: string;
  colors: string[];
  storage?: string[] | null;
  ram?: string[] | null;
  imageColor: string;
  rating: number;
  reviewsCount: number;
  isNew: boolean;
  isBestSeller: boolean;
  isFeatured: boolean;
}

export interface PhoneSpecDto {
  id: string;
  brand: string;
  model: string;
  os: string;
  processor: string;
  ram: string[];
  storage: string[];
  displaySize: string;
  displayType: string;
  resolution: string;
  refreshRate: string;
  brightness: string;
  rearCamera: string;
  frontCamera: string;
  videoRecording: string;
  battery: string;
  charging: string;
  wirelessCharging: boolean;
  reverseCharging: boolean;
  fingerprint: boolean;
  faceUnlock: boolean;
  waterResistance: string;
  weight: string;
  dimensions: string;
  fiveG: boolean;
  nfc: boolean;
  bluetooth: string;
  wifi: string;
  usb: string;
  audioJack: boolean;
  colors: string[];
  priceEGP: number;
  releaseDate: string;
  _batteryMah: number;
  _chargingW: number;
  _refreshRateHz: number;
  _rearMp: number;
  _ramGb: number;
  _displayInch: number;
}

export interface CategoryDto {
  id: string;
  icon: string;
  color: string;
}

export interface ReviewDto {
  id: string;
  storeId?: string | null;
  productId?: string | null;
  author: string;
  authorAr: string;
  rating: number;
  textAr: string;
  textEn: string;
  date: string;
}

export interface DashboardStatsDto {
  storeId: string;
  views: number;
  visitors: number;
  clicks: number;
  reservations: number;
  newReviews: number;
  messagesCount: number;
  saved: number;
  trending: number;
}

export interface OrderDto {
  id: string;
  productAr: string;
  productEn: string;
  storage?: string | null;
  customer: string;
  timeAr: string;
  timeEn: string;
  status: string;
  price: number;
}

export interface MessageDto {
  id: string;
  customer: string;
  productAr: string;
  productEn: string;
  timeAr: string;
  timeEn: string;
  status: string;
  unread: number;
}

export interface CreateProductBody {
  storeId?: string;
  nameAr: string;
  nameEn: string;
  descriptionAr?: string;
  descriptionEn?: string;
  brand: string;
  model: string;
  price: number;
  discountPrice?: number | null;
  category: string;
  condition?: string;
  warranty?: string;
  warrantyAr?: string;
  colors?: string[];
  storage?: string[] | null;
  ram?: string[] | null;
  imageColor?: string;
  inStock?: boolean;
}

export type GetStoresParams = {
search?: string;
governorate?: string;
isVerified?: string;
/**
 * Comma-separated store IDs
 */
ids?: string;
sort?: string;
};

export type GetProductsParams = {
storeId?: string;
category?: string;
isNew?: string;
isBestSeller?: string;
isFeatured?: string;
condition?: string;
/**
 * Comma-separated product IDs
 */
ids?: string;
excludeId?: string;
search?: string;
brand?: string;
};

export type GetDashboardStatsParams = {
storeId?: string;
};

export type GetDashboardOrdersParams = {
storeId?: string;
};

export type GetDashboardMessagesParams = {
storeId?: string;
};

export type GetDashboardReviewsParams = {
storeId?: string;
};