# lib/api-zod/src/generated/types/createProductBody.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * MOB HUB Egypt — Mobile Phone Marketplace API
 * OpenAPI spec version: 0.2.0
 */

export interface CreateProductBody {
  storeId?: string;
  nameAr: string;
  nameEn: string;
  descriptionAr?: string;
  descriptionEn?: string;
  brand: string;
  model: string;
  price: number;
  discountPrice?: number | null;
  category: string;
  condition?: string;
  warranty?: string;
  warrantyAr?: string;
  colors?: string[];
  storage?: string[] | null;
  ram?: string[] | null;
  imageColor?: string;
  inStock?: boolean;
}