# lib/api-zod/src/generated/types/dashboardStatsDto.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * MOB HUB Egypt — Mobile Phone Marketplace API
 * OpenAPI spec version: 0.2.0
 */

export interface DashboardStatsDto {
  storeId: string;
  views: number;
  visitors: number;
  clicks: number;
  reservations: number;
  newReviews: number;
  messagesCount: number;
  saved: number;
  trending: number;
}