# lib/api-zod/src/generated/types/orderDto.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * MOB HUB Egypt — Mobile Phone Marketplace API
 * OpenAPI spec version: 0.2.0
 */

export interface OrderDto {
  id: string;
  productAr: string;
  productEn: string;
  storage?: string | null;
  customer: string;
  timeAr: string;
  timeEn: string;
  status: string;
  price: number;
}