# lib/api-zod/src/generated/types/reviewDto.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * MOB HUB Egypt — Mobile Phone Marketplace API
 * OpenAPI spec version: 0.2.0
 */

export interface ReviewDto {
  id: string;
  storeId?: string | null;
  productId?: string | null;
  author: string;
  authorAr: string;
  rating: number;
  textAr: string;
  textEn: string;
  date: string;
}