# lib/api-zod/src/generated/types/storeDto.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * MOB HUB Egypt — Mobile Phone Marketplace API
 * OpenAPI spec version: 0.2.0
 */

export interface StoreDto {
  id: string;
  nameAr: string;
  nameEn: string;
  descriptionAr: string;
  descriptionEn: string;
  logoColor: string;
  logoInitial: string;
  coverGradient: string[];
  rating: number;
  reviewsCount: number;
  governorate: string;
  city: string;
  address: string;
  addressAr: string;
  lat: number;
  lng: number;
  phone: string;
  whatsapp: string;
  facebook?: string | null;
  instagram?: string | null;
  website?: string | null;
  workingHours: string;
  workingHoursAr: string;
  isOpen: boolean;
  isVerified: boolean;
  productsCount: number;
  categories: string[];
}